+++ src/interfaces/IPackageManager.ts
@@ -7,7 +7,7 @@ import { PackageConfiguration } from "../configuration/models/packages/packageCo
 import { IEnginePipelineStep } from "./IEnginePipelineStep";
 
 export interface IPackageManager extends IEnginePipelineStep {
-    info(logger: ILogger, fileSystem: IFileSystem, packageName: string): Promise<PackageConfiguration>;
+    info(logger: ILogger, fileSystem: IFileSystem, packageName: string, version: string): Promise<PackageConfiguration>;
     add(logger: ILogger, fileSystem: IFileSystem, workingDirectory: string, packageName: string, version: string, isDev: boolean): Promise<void>;
     remove(logger: ILogger, fileSystem: IFileSystem, workingDirectory: string, packageName: string, isDev: boolean): Promise<void>;
 }