+++ src/pipelineSteps/content/htmlTemplate.ts
@@ -58,11 +58,12 @@ export class HtmlTemplate extends EnginePipelineStepBase {
                 const lines: string[] = [];
                 let indent = 0;
                 this.addLine(indent, lines, "<!doctype html>");
-                this.addLine(indent, lines, "<html>");
+                this.addLine(indent, lines, "<html lang=\"en\">");
                 indent++;
                 this.addLine(indent, lines, "<head>");
                 indent++;
                 this.addLine(indent, lines, "<meta charset=\"utf-8\"/>");
+                this.addLine(indent, lines, "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1, shrink-to-fit=no\">");
                 this.addLine(indent, lines, `<title>${uniteConfiguration.title}</title>`);
                 this.addLine(indent, lines, "<link rel=\"stylesheet\" href=\"./css/style.css{CACHEBUST}\">");
                 lines.push("{THEME}");