+++ src/pipelineSteps/packageManager/npm.ts
@@ -20,10 +20,10 @@ export class Npm extends EnginePipelineStepBase implements IPackageManager {
         return 0;
     }
 
-    public async info(logger: ILogger, fileSystem: IFileSystem, packageName: string): Promise<PackageConfiguration> {
+    public async info(logger: ILogger, fileSystem: IFileSystem, packageName: string, version: string): Promise<PackageConfiguration> {
         logger.info("Looking up package info...");
 
-        const args = ["view", packageName, "--json", "name", "version", "main"];
+        const args = ["view", `${packageName}${version !== null && version !== undefined ? `@${version}` : ""}`, "--json", "name", "version", "main"];
 
         return PackageUtils.exec(logger, fileSystem, "npm", undefined, args)
             .then(viewData => JSON.parse(viewData))